+++ components/examples/examples-index-page.js
@@ -1,0 +1,15 @@
+import Feedback from '../docs/feedback';
+import DocsLayout from '../docs/docs-layout';
+
+function ExamplesIndexPage() {
+  return (
+    <DocsLayout>
+      <h1>Examples</h1>
+      <p>(Intro paragraph)</p>
+      <hr />
+      <Feedback />
+    </DocsLayout>
+  );
+}
+
+export default ExamplesIndexPage;
